feat: add album and artist genre retrieval features
The update includes new features to the SpotifyAPI.ts interface, adding functionality to pull album and artist data by ID. Complementary updates to the Player.vue component reflect the consumption of this new data, displaying album and artist genre information directly on the player. In addition to providing more information for the user, this commit assists in streamlining album and artist genre-related operations.